+++ src/Credfeto.ChangeLog.Cmd/Options.cs
@@ -27,5 +27,8 @@ namespace Credfeto.ChangeLog.Cmd
 
         [Option(shortName: 'p', longName: "Pending", SetName = "Release", Required = false, HelpText = "If the release is pending")]
         public bool Pending { get; set; }
+
+        [Option(shortName: 'u', longName: "UnReleased", SetName = "Extract", Required = false, HelpText = "Prints the unreleased section to the console.")]
+        public bool DisplayUnreleased { get; init; }
     }
 }